format code

+++ src/Token.sol
@@ -8,19 +8,18 @@ import "@openzeppelin/contracts/token/ERC20/extensions/ERC20Pausable.sol";
 /**
  * @title MyToken
  * @dev A customizable ERC20 token with role-based access control, blacklist functionality, pausable features, and transfer fees.
- * 
+ *
  * This contract extends OpenZeppelin's ERC20, ERC20Pausable, and AccessControl to provide:
  * - Standard ERC20 token functionality
  * - Role-based access control for different operations
  * - Blacklist management to prevent transfers from/to specific addresses
  * - Pausable functionality to halt all token transfers
  * - Transfer fee mechanism with configurable percentage
- * 
+ *
  * @custom:security This contract uses AccessControl for secure role management.
  * Only authorized addresses can perform sensitive operations like minting, burning, and blacklist management.
  */
 contract MyToken is ERC20, ERC20Pausable, AccessControl {
-
     // Define custom roles
     bytes32 public constant MINTER_ROLE = keccak256("MINTER_ROLE");
     bytes32 public constant BURNER_ROLE = keccak256("BURNER_ROLE");
@@ -31,16 +30,16 @@ contract MyToken is ERC20, ERC20Pausable, AccessControl {
 
     /// @dev Mapping to track blacklisted addresses
     mapping(address => bool) blacklist;
-    
+
     /// @dev Transfer fee percentage (in basis points, 100 = 1%)
     uint256 public transferFeePercentage;
-    
+
     /// @dev Address that receives the transfer fees
     address public feeCollector;
-    
+
     /// @dev Maximum fee percentage (10% = 1000 basis points)
     uint256 public constant MAX_FEE_PERCENTAGE = 1000;
-    
+
     /// @dev Basis points denominator (100% = 10000)
     uint256 public constant BASIS_POINTS = 10000;
 
@@ -86,8 +85,10 @@ contract MyToken is ERC20, ERC20Pausable, AccessControl {
      * @param feeAmount The fee amount collected
      * @param netAmount The net amount received by recipient
      */
-    event TransferFeeCollected(address indexed from, address indexed to, uint256 amount, uint256 feeAmount, uint256 netAmount);
-    
+    event TransferFeeCollected(
+        address indexed from, address indexed to, uint256 amount, uint256 feeAmount, uint256 netAmount
+    );
+
     //============================//
     //           Constructor      //
     //============================//
@@ -98,24 +99,21 @@ contract MyToken is ERC20, ERC20Pausable, AccessControl {
      * @param symbol The symbol of the token
      * @param initialFeePercentage The initial transfer fee percentage (in basis points)
      * @param initialFeeCollector The initial address to receive transfer fees
-     * 
+     *
      * Sets up the initial roles:
      * - Deployer gets DEFAULT_ADMIN_ROLE
      * - Deployer gets all custom roles initially
      * - Mints initial supply to deployer
      */
-    constructor(
-        string memory name, 
-        string memory symbol, 
-        uint256 initialFeePercentage, 
-        address initialFeeCollector
-    ) ERC20(name, symbol) {
+    constructor(string memory name, string memory symbol, uint256 initialFeePercentage, address initialFeeCollector)
+        ERC20(name, symbol)
+    {
         require(initialFeePercentage <= MAX_FEE_PERCENTAGE, "Fee percentage too high");
         require(initialFeeCollector != address(0), "Invalid fee collector");
-        
+
         // Grant the contract deployer the default admin role
         _grantRole(DEFAULT_ADMIN_ROLE, msg.sender);
-        
+
         // Grant the deployer all custom roles initially
         _grantRole(MINTER_ROLE, msg.sender);
         _grantRole(BURNER_ROLE, msg.sender);
@@ -123,11 +121,11 @@ contract MyToken is ERC20, ERC20Pausable, AccessControl {
         _grantRole(PAUSER_ROLE, msg.sender);
         _grantRole(ADMIN_ROLE, msg.sender);
         _grantRole(FEE_MANAGER_ROLE, msg.sender);
-        
+
         // Set initial fee parameters
         transferFeePercentage = initialFeePercentage;
         feeCollector = initialFeeCollector;
-        
+
         _mint(msg.sender, 1000 * 10 ** decimals());
     }
 
@@ -138,13 +136,13 @@ contract MyToken is ERC20, ERC20Pausable, AccessControl {
     /**
      * @dev Mints tokens to the caller's address.
      * @param amount The amount of tokens to mint
-     * 
+     *
      * Requirements:
      * - Caller must have MINTER_ROLE
-     * 
+     *
      * Emits a {Transfer} event with `from` set to the zero address.
      */
-    function mint(uint amount) external onlyRole(MINTER_ROLE) {
+    function mint(uint256 amount) external onlyRole(MINTER_ROLE) {
         _mint(msg.sender, amount);
     }
 
@@ -152,27 +150,27 @@ contract MyToken is ERC20, ERC20Pausable, AccessControl {
      * @dev Mints tokens to a specific address.
      * @param to The address to mint tokens to
      * @param amount The amount of tokens to mint
-     * 
+     *
      * Requirements:
      * - Caller must have MINTER_ROLE
-     * 
+     *
      * Emits a {Transfer} event with `from` set to the zero address.
      */
-    function mintTo(address to, uint amount) external onlyRole(MINTER_ROLE) {
+    function mintTo(address to, uint256 amount) external onlyRole(MINTER_ROLE) {
         _mint(to, amount);
     }
 
     /**
      * @dev Burns tokens from the caller's address.
      * @param amount The amount of tokens to burn
-     * 
+     *
      * Requirements:
      * - Caller must have BURNER_ROLE
      * - Caller must have sufficient balance
-     * 
+     *
      * Emits a {Transfer} event with `to` set to the zero address.
      */
-    function burn(uint amount) external onlyRole(BURNER_ROLE) {
+    function burn(uint256 amount) external onlyRole(BURNER_ROLE) {
         _burn(msg.sender, amount);
     }
 
@@ -180,25 +178,25 @@ contract MyToken is ERC20, ERC20Pausable, AccessControl {
      * @dev Burns tokens from a specific address.
      * @param from The address to burn tokens from
      * @param amount The amount of tokens to burn
-     * 
+     *
      * Requirements:
      * - Caller must have BURNER_ROLE
      * - `from` must have sufficient balance
-     * 
+     *
      * Emits a {Transfer} event with `to` set to the zero address.
      */
-    function burnFrom(address from, uint amount) external onlyRole(BURNER_ROLE) {
+    function burnFrom(address from, uint256 amount) external onlyRole(BURNER_ROLE) {
         _burn(from, amount);
     }
 
     /**
      * @dev Adds an address to the blacklist, preventing transfers from/to that address.
      * @param user The address to blacklist
-     * 
+     *
      * Requirements:
      * - Caller must have BLACKLIST_MANAGER_ROLE
      * - User must not already be blacklisted
-     * 
+     *
      * Emits a {UserBlacklisted} event.
      */
     function addToBlacklist(address user) external onlyRole(BLACKLIST_MANAGER_ROLE) {
@@ -210,11 +208,11 @@ contract MyToken is ERC20, ERC20Pausable, AccessControl {
     /**
      * @dev Removes an address from the blacklist, allowing transfers from/to that address.
      * @param user The address to remove from blacklist
-     * 
+     *
      * Requirements:
      * - Caller must have BLACKLIST_MANAGER_ROLE
      * - User must be currently blacklisted
-     * 
+     *
      * Emits a {UserRemovedFromBlacklist} event.
      */
     function removeFromBlacklist(address user) external onlyRole(BLACKLIST_MANAGER_ROLE) {
@@ -225,10 +223,10 @@ contract MyToken is ERC20, ERC20Pausable, AccessControl {
 
     /**
      * @dev Pauses all token transfers.
-     * 
+     *
      * Requirements:
      * - Caller must have PAUSER_ROLE
-     * 
+     *
      * Emits a {Paused} event.
      */
     function pause() public onlyRole(PAUSER_ROLE) {
@@ -237,10 +235,10 @@ contract MyToken is ERC20, ERC20Pausable, AccessControl {
 
     /**
      * @dev Unpauses all token transfers.
-     * 
+     *
      * Requirements:
      * - Caller must have PAUSER_ROLE
-     * 
+     *
      * Emits an {Unpaused} event.
      */
     function unpause() public onlyRole(PAUSER_ROLE) {
@@ -252,11 +250,11 @@ contract MyToken is ERC20, ERC20Pausable, AccessControl {
     /**
      * @dev Updates the transfer fee percentage.
      * @param newFeePercentage The new fee percentage (in basis points)
-     * 
+     *
      * Requirements:
      * - Caller must have FEE_MANAGER_ROLE
      * - New fee percentage must not exceed MAX_FEE_PERCENTAGE
-     * 
+     *
      * Emits a {TransferFeeUpdated} event.
      */
     function setTransferFeePercentage(uint256 newFeePercentage) external onlyRole(FEE_MANAGER_ROLE) {
@@ -269,11 +267,11 @@ contract MyToken is ERC20, ERC20Pausable, AccessControl {
     /**
      * @dev Updates the fee collector address.
      * @param newFeeCollector The new address to receive transfer fees
-     * 
+     *
      * Requirements:
      * - Caller must have FEE_MANAGER_ROLE
      * - New fee collector must not be zero address
-     * 
+     *
      * Emits a {FeeCollectorUpdated} event.
      */
     function setFeeCollector(address newFeeCollector) external onlyRole(FEE_MANAGER_ROLE) {
@@ -312,7 +310,7 @@ contract MyToken is ERC20, ERC20Pausable, AccessControl {
     /**
      * @dev Grants MINTER_ROLE to an account.
      * @param account The address to grant the role to
-     * 
+     *
      * Requirements:
      * - Caller must have ADMIN_ROLE
      */
@@ -323,7 +321,7 @@ contract MyToken is ERC20, ERC20Pausable, AccessControl {
     /**
      * @dev Revokes MINTER_ROLE from an account.
      * @param account The address to revoke the role from
-     * 
+     *
      * Requirements:
      * - Caller must have ADMIN_ROLE
      */
@@ -334,7 +332,7 @@ contract MyToken is ERC20, ERC20Pausable, AccessControl {
     /**
      * @dev Grants BURNER_ROLE to an account.
      * @param account The address to grant the role to
-     * 
+     *
      * Requirements:
      * - Caller must have ADMIN_ROLE
      */
@@ -345,7 +343,7 @@ contract MyToken is ERC20, ERC20Pausable, AccessControl {
     /**
      * @dev Revokes BURNER_ROLE from an account.
      * @param account The address to revoke the role from
-     * 
+     *
      * Requirements:
      * - Caller must have ADMIN_ROLE
      */
@@ -356,7 +354,7 @@ contract MyToken is ERC20, ERC20Pausable, AccessControl {
     /**
      * @dev Grants BLACKLIST_MANAGER_ROLE to an account.
      * @param account The address to grant the role to
-     * 
+     *
      * Requirements:
      * - Caller must have ADMIN_ROLE
      */
@@ -367,7 +365,7 @@ contract MyToken is ERC20, ERC20Pausable, AccessControl {
     /**
      * @dev Revokes BLACKLIST_MANAGER_ROLE from an account.
      * @param account The address to revoke the role from
-     * 
+     *
      * Requirements:
      * - Caller must have ADMIN_ROLE
      */
@@ -378,7 +376,7 @@ contract MyToken is ERC20, ERC20Pausable, AccessControl {
     /**
      * @dev Grants PAUSER_ROLE to an account.
      * @param account The address to grant the role to
-     * 
+     *
      * Requirements:
      * - Caller must have ADMIN_ROLE
      */
@@ -389,7 +387,7 @@ contract MyToken is ERC20, ERC20Pausable, AccessControl {
     /**
      * @dev Revokes PAUSER_ROLE from an account.
      * @param account The address to revoke the role from
-     * 
+     *
      * Requirements:
      * - Caller must have ADMIN_ROLE
      */
@@ -400,7 +398,7 @@ contract MyToken is ERC20, ERC20Pausable, AccessControl {
     /**
      * @dev Grants FEE_MANAGER_ROLE to an account.
      * @param account The address to grant the role to
-     * 
+     *
      * Requirements:
      * - Caller must have ADMIN_ROLE
      */
@@ -411,7 +409,7 @@ contract MyToken is ERC20, ERC20Pausable, AccessControl {
     /**
      * @dev Revokes FEE_MANAGER_ROLE from an account.
      * @param account The address to revoke the role from
-     * 
+     *
      * Requirements:
      * - Caller must have ADMIN_ROLE
      */
@@ -422,7 +420,7 @@ contract MyToken is ERC20, ERC20Pausable, AccessControl {
     /**
      * @dev Grants ADMIN_ROLE to an account.
      * @param account The address to grant the role to
-     * 
+     *
      * Requirements:
      * - Caller must have DEFAULT_ADMIN_ROLE
      */
@@ -433,7 +431,7 @@ contract MyToken is ERC20, ERC20Pausable, AccessControl {
     /**
      * @dev Revokes ADMIN_ROLE from an account.
      * @param account The address to revoke the role from
-     * 
+     *
      * Requirements:
      * - Caller must have DEFAULT_ADMIN_ROLE
      */
@@ -455,12 +453,12 @@ contract MyToken is ERC20, ERC20Pausable, AccessControl {
      * @param from The address tokens are transferred from
      * @param to The address tokens are transferred to
      * @param value The amount of tokens being transferred
-     * 
+     *
      * Requirements:
      * - `from` must not be blacklisted
      * - `to` must not be blacklisted
      * - Contract must not be paused
-     * 
+     *
      * If transfer fee is enabled:
      * - Calculates fee amount
      * - Transfers fee to fee collector
@@ -468,19 +466,19 @@ contract MyToken is ERC20, ERC20Pausable, AccessControl {
      * - Emits {TransferFeeCollected} event
      */
     function _update(address from, address to, uint256 value) internal override(ERC20, ERC20Pausable) {
-        require(!blacklist[from], 'Sender is blacklisted');
-        require(!blacklist[to], 'Recipient is blacklisted');
-        
+        require(!blacklist[from], "Sender is blacklisted");
+        require(!blacklist[to], "Recipient is blacklisted");
+
         // Handle transfer fee logic
         if (transferFeePercentage > 0 && from != address(0) && to != address(0)) {
             (uint256 feeAmount, uint256 netAmount) = calculateTransferFee(value);
-            
+
             if (feeAmount > 0) {
                 // Transfer fee to collector
                 super._update(from, feeCollector, feeAmount);
                 // Transfer net amount to recipient
                 super._update(from, to, netAmount);
-                
+
                 emit TransferFeeCollected(from, to, value, feeAmount, netAmount);
             } else {
                 // No fee, normal transfer
